# Backend/labfastorder-openapi.yaml
openapi: 3.1.0
info:
  title: "labfastorder API"
  description: "labfastorder API"
  version: "1.0.0"
servers:
  - url: "localhost:8080"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        username:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        roles:
          type: "array"
          items:
            $ref: "#/components/schemas/Role"
    Role:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        name:
          type: "string"
          enum:
            - "ROLE_USER"
            - "ROLE_ADMIN"
    PasswordChange:
      type: "object"
      properties:
        oldPassword:
          type: "string"
        newPassword:
          type: "string"
security:
  - bearerAuth: []
paths:
...
  /api/user/all:
    get:
      summary: "GET api/user/all"
      operationId: "getAllUsers"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/User"
  /api/user/delete/{id}:
    delete:
      summary: "DELETE api/user/delete/{id}"
      operationId: "deleteUser"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /api/user/changePassword/{id}:
    put:
      summary: "PUT api/user/changePassword/{id}"
      operationId: "changePassword"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChange"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /api/user/changeRole/{id}:
    put:
      summary: "PUT api/user/changeRole/{id}"
      operationId: "changeRole"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              type: "string"
              enum:
                - "ROLE_USER"
                - "ROLE_ADMIN"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /api/user/changeUsername/{id}:
    put:
      summary: "PUT api/user/changeUsername/{id}"
      operationId: "changeUsername"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              type: "string"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"